Chore : queryDSL 실습/week6

+++ 7th/src/main/java/UMC/_th/domain/Review.java
@@ -18,7 +18,7 @@ public class Review extends BaseEntity {
     private String title;
 
     @Column(nullable = false)
-    private Integer starPoint;
+    private Float starPoint;
 
     @Column(length = 256)
     private String description;
